@@ -11,22 +11,43 @@ import (
 	"github.com/chromedp/chromedp"
 )
 
-var (
-	FailedLoadReloadDelay = time.Second * 10
-)
+var FailedLoadReloadDelay = time.Second * 10
 
 type Tab struct {
-	Browser Browser
+	Browser *Browser
 	Context context.Context
 
 	Close func()
 
 	isReloading bool
+
+	css string
+	js  string
 }
 
 type BasicAuthCredentials struct {
 	Username string
 	Password string
+}
+
+func newTab(b *Browser) *Tab {
+	ctx, close := chromedp.NewContext(b.Context)
+
+	t := &Tab{
+		Browser: b,
+		Context: ctx,
+		Close:   close,
+	}
+
+	chromedp.ListenTarget(ctx, func(ev interface{}) {
+		// When the page is done loading, inject the CSS and JS
+		if _, ok := ev.(*page.EventFrameStoppedLoading); ok {
+			go t.injectCSS()
+			go t.injectJS()
+		}
+	})
+
+	return t
 }
 
 func (t *Tab) Focus() {
@@ -53,12 +74,17 @@ func (t *Tab) NavigateWithBasicAuth(url string, creds BasicAuthCredentials) {
 	)
 }
 
-func (t *Tab) AddCSS(css string) {
+func (t *Tab) SetCSS(css string) {
+	t.css = css
+	t.injectCSS()
+}
+
+func (t *Tab) injectCSS() {
 	script := `
 	(() => {
 		const style = document.createElement('style');
 		style.type = 'text/css';
-		style.appendChild(document.createTextNode(` + "`" + css + "`" + `));
+		style.appendChild(document.createTextNode(` + "`" + t.css + "`" + `));
 		document.head.appendChild(style);
 	})()
 	`
@@ -69,11 +95,16 @@ func (t *Tab) AddCSS(css string) {
 	)
 }
 
-func (t *Tab) AddJS(script string) {
+func (t *Tab) SetJS(js string) {
+	t.js = js
+	t.injectJS()
+}
+
+func (t *Tab) injectJS() {
 	var executed *runtime.RemoteObject
 	chromedp.Run(
 		t.Context,
-		chromedp.EvaluateAsDevTools(script, &executed),
+		chromedp.EvaluateAsDevTools(t.js, &executed),
 	)
 }
 
